requirements.txt
beautifulsoup4==4.12.3
colorama==0.4.6
Distance==0.1.3
Flask==3.0.3
python-slugify==8.0.4
requests==2.32.3
